Main/attmain/requirements.txt
wheel
click
RPi.GPIO
PyFingerprint
